+++ app/components/bs-alert.js
@@ -1,4 +1,4 @@
 import Ember from 'ember';
-import component from 'ember-bootstrap/components/bs-alert';
+import component from 'ember-bootstrap-components/components/bs-alert';
 
 export default component;