+++ app/components/bs-button-group.js
@@ -1,4 +1,4 @@
 import Ember from 'ember';
-import bsButtonGroup from 'ember-bootstrap/components/bs-button-group';
+import bsButtonGroup from 'ember-bootstrap-components/components/bs-button-group';
 
 export default bsButtonGroup;