+++ app/components/bs-button.js
@@ -1,4 +1,4 @@
 import Ember from 'ember';
-import bsButton from 'ember-bootstrap/components/bs-button';
+import bsButton from 'ember-bootstrap-components/components/bs-button';
 
 export default bsButton;