+++ app/components/bs-form-group.js
@@ -1,4 +1,4 @@
 import Ember from 'ember';
-import component from 'ember-bootstrap/components/bs-form-group';
+import component from 'ember-bootstrap-components/components/bs-form-group';
 
 export default component;