+++ app/components/bs-input.js
@@ -1,4 +1,4 @@
 import Ember from 'ember';
-import component from 'ember-bootstrap/components/bs-input';
+import component from 'ember-bootstrap-components/components/bs-input';
 
 export default component;