+++ app/helpers/add-waves.js
@@ -1,0 +1,1 @@
+export { default, addWaves } from 'ember-bootstrap-components/helpers/add-waves';