+++ app/instance-initializers/register-modal.js
@@ -1,0 +1,7 @@
+export default {
+  name: 'register-modal',
+  initialize: function(instance) {
+	window.Bootstrap = {};
+	window.Bootstrap.ModalManager = instance.container.lookup("service:bootstrap-modal-manager");
+  }
+};