+++ blueprints/ember-bootstrap/index.js
@@ -1,8 +1,0 @@
-module.exports = {
-    normalizeEntityName: function() {
-    },
-
-    afterInstall: function() {
-        return this.addBowerPackageToProject('bootstrap', '~3.3.2');
-    }
-};