+++ tests/unit/helpers/add-waves-test.js
@@ -1,0 +1,10 @@
+import { waves } from '../../../helpers/add-waves';
+import { module, test } from 'qunit';
+
+module('Unit | Helper | waves');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  var result = waves(42);
+  assert.ok(result);
+});